Fix crackling/buzzing artifacts on bass-heavy N64 USF tracks
Three independent bugs combined to produce distortion on tracks with
loud bass and large ADPCM scale factors:

1. ADDMIXER writeback missing (audio_ucode2.cpp, HLE path):
   ADDMIXER computed the clamped sum of two sample buffers into `temp`
   but never stored it back via `*outp`. The mixed result was silently
   discarded every iteration, leaving the output buffer unchanged.
   Fix: add `*outp = (s16)temp` before advancing the pointer.

2. ADPCM accumulator int32 overflow (audio_ucode1.c):
   The ADPCM predictor accumulates up to 10 products of two 16-bit
   values into `int a[8]`. At large scale factors the running total can
   exceed INT32_MAX (~2.1e9 vs a possible ~1.07e10), wrapping around and
   producing large-amplitude spikes — the characteristic crackling on
   bass-heavy tracks. Fix: widen the accumulator to int64_t; clamp back
   to int16 on output with explicit (short) casts.

3. 3-tap median filter + corrected vol loop (audio.c):
   The volume/byte-swap loop was rewritten to correctly reflect the
   RDRAM little-endian layout produced by the RSP DMA (memcpy without
   XOR byte-swapping): input is buf[i]=lo, buf[i+1]=hi; output is
   big-endian for AU/FLAC/playback. A 3-tap median filter is applied
   per stereo channel after the vol loop to suppress any remaining
   single-sample impulse spikes without attenuating normal audio content.

+++ audio_ucode1.c
@@ -496,7 +496,7 @@ static void ADPCM ()   // Work in progress! :)
     int vscale;
     unsigned short index;
     unsigned short j;
-    int a[8];
+    int64_t a[8];
     short *book1,*book2;
     int l1,l2;
     int inp1[8];
@@ -660,7 +660,7 @@ static void ADPCM ()   // Work in progress! :)
             {
                 a[j^1]=-32768;
             }
-            *(out++)=a[j^1];
+            *(out++)=(short)a[j^1];
         }
         l1=a[6];
         l2=a[7];
@@ -736,7 +736,7 @@ static void ADPCM ()   // Work in progress! :)
             {
                 a[j^1]=-32768;
             }
-            *(out++)=a[j^1];
+            *(out++)=(short)a[j^1];
         }
         l1=a[6];
         l2=a[7];
